# backend/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Tingyue Backend API",
    "version": "1.0.0",
    "description": "Authoritative API contract api-contract-v1.0.0. JSON fields use camelCase and timestamps use UTC RFC 3339. Clients branch on stable error.code values, never human messages."
  },
  "servers": [
    { "url": "http://127.0.0.1:3100", "description": "Local development" },
    { "url": "https://api.test.example.invalid", "description": "Test placeholder; replace after domain approval" }
  ],
...
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/api/v1/works": {
      "get": {
        "operationId": "listWorks",
        "tags": ["Content"],
        "parameters": [
          { "$ref": "#/components/parameters/Cursor" },
          { "$ref": "#/components/parameters/Limit" }
        ],
        "responses": {
          "200": {
            "description": "Visible works in stable cursor order.",
            "headers": { "X-Request-Id": { "$ref": "#/components/headers/RequestId" } },
            "content": { "application/json": { "schema": { "$ref": "./schemas/api-v1.schema.json#/$defs/WorksResponse" } } }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/api/v1/works/{workId}": {
      "get": {
        "operationId": "getWork",
        "tags": ["Content"],
        "parameters": [{ "$ref": "#/components/parameters/WorkId" }],
        "responses": {
          "200": {
            "description": "Work and piece metadata.",
            "headers": { "X-Request-Id": { "$ref": "#/components/headers/RequestId" } },
            "content": { "application/json": { "schema": { "$ref": "./schemas/api-v1.schema.json#/$defs/WorkResponse" } } }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/api/v1/pieces/{pieceId}/manifest": {
      "get": {
        "operationId": "getPieceManifest",
        "tags": ["Content"],
        "parameters": [
          { "$ref": "#/components/parameters/PieceId" },
          {
            "name": "contentVersion",
            "in": "query",
            "required": false,
            "description": "Omit for the current published version.",
            "schema": { "type": "integer", "minimum": 1 }
          }
        ],
        "responses": {
          "200": {
            "description": "Immutable versioned manifest with short-lived asset URLs and SHA-256 checksums.",
            "headers": { "X-Request-Id": { "$ref": "#/components/headers/RequestId" } },
            "content": { "application/json": { "schema": { "$ref": "./schemas/api-v1.schema.json#/$defs/ManifestResponse" } } }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Short-lived access token. Tokens must not be logged."
      }
    },
    "headers": {
      "RequestId": {
        "description": "Opaque request correlation identifier.",
        "required": true,
        "schema": { "type": "string", "minLength": 8, "maxLength": 128 }
      }
    },
    "parameters": {
...
      "Cursor": {
        "name": "cursor",
        "in": "query",
        "required": false,
        "description": "Opaque cursor returned by the preceding page.",
        "schema": { "type": "string", "minLength": 1, "maxLength": 2048 }
      },
      "Limit": {
        "name": "limit",
        "in": "query",
        "required": false,
        "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 20 }
      },
      "WorkId": { "name": "workId", "in": "path", "required": true, "schema": { "type": "string", "minLength": 1, "maxLength": 128 } },
      "PieceId": { "name": "pieceId", "in": "path", "required": true, "schema": { "type": "string", "minLength": 1, "maxLength": 128 } },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Malformed request, invalid ranking period or invalid filter.",
        "headers": { "X-Request-Id": { "$ref": "#/components/headers/RequestId" } },
        "content": { "application/json": { "schema": { "$ref": "./schemas/api-v1.schema.json#/$defs/ErrorResponse" } } }
      },
      "Unauthorized": {
        "description": "Authentication failed, expired or was revoked.",
        "headers": { "X-Request-Id": { "$ref": "#/components/headers/RequestId" } },
        "content": { "application/json": { "schema": { "$ref": "./schemas/api-v1.schema.json#/$defs/ErrorResponse" } } }
      },
      "Forbidden": {
        "description": "Authenticated identity lacks the required permission or content access.",
        "headers": { "X-Request-Id": { "$ref": "#/components/headers/RequestId" } },
        "content": { "application/json": { "schema": { "$ref": "./schemas/api-v1.schema.json#/$defs/ErrorResponse" } } }
      },
      "NotFound": {
        "description": "Business resource not found or not visible.",
        "headers": { "X-Request-Id": { "$ref": "#/components/headers/RequestId" } },
        "content": { "application/json": { "schema": { "$ref": "./schemas/api-v1.schema.json#/$defs/ErrorResponse" } } }
      },
      "Conflict": {
        "description": "State, version or idempotency conflict.",
        "headers": { "X-Request-Id": { "$ref": "#/components/headers/RequestId" } },
        "content": { "application/json": { "schema": { "$ref": "./schemas/api-v1.schema.json#/$defs/ErrorResponse" } } }
      },
...
    }
  }
}